requirements.txt
Flask>=2.1.0
tensorflow>=2.5.0
Pillow>=8.2.0
scikit-learn>=0.24.0
gunicorn>=20.1.0
